% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AURN_O3}
\alias{AURN_O3}
\title{Automatic Urban and Rural Network (AURN) O\if{html}{\out{<sub>}}3\if{html}{\out{</sub>}} data.}
\format{
\subsection{\code{AURN_O3}}{

A data frame with 6390 rows and 9 columns:
\describe{
\item{U}{Longitude of the station location (in -180 to 180.0 format)}
\item{V}{Latitude of the station location (in -90.0 to 90.0 format)}
\item{T}{Time index of measurement}
\item{tstamp}{Time stamp of measurement (in YYY-MM-DD format)}
\item{SID}{Site ID of measurement location}
\item{SCODE}{Site Code}
\item{SFNAME}{Site Name}
\item{STYPE}{Site type - will be one of Urban Background, Rural Background, Urban Traffic, Urban Industrial, Suburban Industrial or Suburban Background}
\item{Poll_Value}{Daily concentration of pollutant O\if{html}{\out{<sub>}}3\if{html}{\out{</sub>}} in \eqn{\mu} gm\if{html}{\out{<sup>}}-3\if{html}{\out{</sup>}}}
}
}
}
\source{
\url{https://uk-air.defra.gov.uk/data/data_selector}
}
\usage{
AURN_O3
}
\description{
This dataset contains daily aggregated air quality data from the United Kingdom's Automatic Urban
and Rural Network (AURN) for the secondary air pollutant ozone (O\if{html}{\out{<sub>}}3\if{html}{\out{</sub>}}).Data is provided for
30 sites around the UK that provided a complete daily mean dataset (with no missing data) for the time period
1\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} February to 31\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} August 2020. The data is already formatted into a data frame for
running with GAM-PELT.

A description of the AURN network is available from the \href{https://uk-air.defra.gov.uk/networks/network-info?view=aurn}{UK-AIR}
website.The data are freely available to download and are provided by the Department for the Environment and Rural Affairs
(Defra) under the following attribution statement:

(C) Crown copyright 2021 Defra via uk-air.defra.gov.uk, licensed under the \href{https://www.nationalarchives.gov.uk/doc/open-government-licence/version/2/]}{Open Government Licence}.
}
\keyword{datasets}
