% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_input.R
\name{class_input}
\alias{class_input}
\title{Class Input Function}
\usage{
class_input(data, pen.type, pen.value, minseglen, cpts, GAM.form, GAM_call)
}
\arguments{
\item{data}{Data used in the changepoint analysis, see \code{\link{GAM.PELT}} for further details.}

\item{pen.type}{Penalty used as a text string, see \code{\link{GAM.PELT}} for further details.}

\item{pen.value}{Numerical value of the penalty used in the analysis (positive).}

\item{minseglen}{Minimum segment length used in the analysis (positive integer).}

\item{cpts}{Locations of changepoints identified by the analysis.}

\item{GAM.form}{GAM formula used in the changepoint analysis, see \code{\link{GAM.PELT}} for further details.}

\item{GAM_call}{Call to fit the GAM model used in the changepoint analysis. Internal variable used to output
the fitted GAM parameters for each segment based on the changepoint locations.}
}
\value{
An object of class \code{GAM_PELT} filled with the given attributes.
}
\description{
This function helps to input all the necessary information into the correct format for
\code{GAM_PELT} class. This function is called by \code{GAM.PELT}. This is not intended for use
by regular users of the package.

WARNING: No checks on arguments are performed!
}
\details{
This function takes all the input required for the \code{GAM_PELT} class and enters it into the object.
This function is exported for developer use only.
}
\author{
Michael Hollaway

Rebecca Killick
}
