% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_function.R
\name{cost_function}
\alias{cost_function}
\title{Cost function for fitting GAM to current segment}
\usage{
cost_function(tloc_1, tloc_2, df_in, GAM_call, minseglen)
}
\arguments{
\item{tloc_1}{Time index of start of segment}

\item{tloc_2}{Time index of end of segment}

\item{df_in}{Input data frame to fit GAM to}

\item{GAM_call}{Object containing all the parameters of the GAM to evaluate.}

\item{minseglen}{Minimum segment length for GAM fit (still needed?)}
}
\value{
Cost of fitting the GAM to the segment indicated by tloc_1 and tloc_2
}
\description{
Cost function for fitting GAM to current segment
}
\author{
Michael Hollaway

Rebecca Killick
}
