% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gampelt.R
\docType{methods}
\name{cpts}
\alias{cpts}
\alias{cpts-methods}
\alias{cpts,GAM_PELT-method}
\alias{cpts,GAM_PELT,cpts-methods}
\title{Generic Function - cpts}
\usage{
cpts(object)

\S4method{cpts}{GAM_PELT}(object)
}
\arguments{
\item{object}{Object of class \code{GAM_PELT}}
}
\value{
Returns an ordered vector of the optimal number of changepoints identified by the call to \code{GAM.PELT}
}
\description{
Generic Function
}
\details{
Generic Function
}
\examples{
#Generate a new object of class GAM_PELT
x <- methods::new('GAM_PELT')
#Return the vector of changepoints - will be blank in new object.
cpts(x)

}
\author{
Michael Hollaway

Rebecca Killick
}
