% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gampelt.R
\docType{methods}
\name{data.set}
\alias{data.set}
\alias{data.set-methods}
\alias{data.set,GAM_PELT-method}
\alias{data.set,GAM_PELT,data.set-methods}
\title{Generic Function - data.set}
\usage{
data.set(object)

\S4method{data.set}{GAM_PELT}(object)
}
\arguments{
\item{object}{Object of class \code{GAM_PELT}}
}
\value{
A data frame containing the original dataset.
}
\description{
Generic Function
}
\details{
Generic Function
}
\examples{
#Generate a new object of class GAM_PELT
x <- methods::new('GAM_PELT')
#Retrieve the data.set slot from x.
data.set(x)

}
\author{
Michael Hollaway

Rebecca Killick
}
