% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gampelt.R
\name{GAM.form}
\alias{GAM.form}
\alias{GAM.form-methods}
\alias{GAM.form,GAM_PELT-method}
\alias{GAM.form,GAM_PELT,GAM.form-methods}
\title{Generic Function - GAM.form}
\usage{
GAM.form(object)

\S4method{GAM.form}{GAM_PELT}(object)
}
\arguments{
\item{object}{Object of class \code{GAM_PELT}}
}
\value{
The GAM formula used in the analysis.
}
\description{
Generic Function
}
\details{
Generic Function
}
\examples{
#Generate a new object of class GAM_PELT
x <- methods::new('GAM_PELT')
#Retrieve the GAM.form slot from x.
GAM.form(x)

}
\author{
Michael Hollaway

Rebecca Killick
}
