% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gampelt.R
\docType{class}
\name{GAM_PELT-class}
\alias{GAM_PELT-class}
\title{GAM_PELT-class}
\description{
An S4 class to store the output of GAM.PELT
}
\section{Slots}{

\describe{
\item{\code{data.set}}{Object of class \code{"data.frame"}, the original dataset.}

\item{\code{pen.type}}{Object of class \code{"character"}, the penalty type specified in the analysis.}

\item{\code{pen.value}}{Object of class \code{"numeric"}, the value of the penalty used in the analysis.}

\item{\code{minseglen}}{Object of class \code{"numeric"}, the minimum segment length (number of observations between changepoints) used in the analysis.}

\item{\code{cpts}}{Object of class \code{"numeric"}, vector of changepoints identified.}

\item{\code{GAM.form}}{Object of class \code{"formula"}, the specified GAM formula.}

\item{\code{param.est}}{Object of class \code{"list"}, a list where each element is a vector of GAM model parameter estimates.}

\item{\code{ncpts}}{Object of class \code{"numeric"}, the total number of cpts identified.}
}}

