% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimStudies_gampelt.R
\name{gen_SimData}
\alias{gen_SimData}
\title{Generate a test simulation dataset.}
\usage{
gen_SimData(
  SCENARIO,
  seed,
  n_ts = 500,
  n_sites = 50,
  true_cpts = c(100, 200, 300, 400)
)
}
\arguments{
\item{SCENARIO}{Chosen scenario, will be one of either 1a, 1b, 1c, 2a, 2b, 2c (temporal changes only),
3a, 3b, 3c (spatial changes only), 4a, 4b (spatio-temporal changes), A, B, C, D, E, F (no changes).}

\item{seed}{Random seed to use for reproducibility.}

\item{n_ts}{Number of timesteps in dataset. Default is 500.}

\item{n_sites}{Number of spatial locations in dataset. Default is 50}

\item{true_cpts}{Locations of the true changepoints in the dataset. Default is 100, 200, 300, and 400 when n_ts=500.}
}
\value{
A list with the elements:

\code{data} - A data frame containing the data in the correct format to run \code{GAM-PELT}. This contains
the model response variable (Y) and covariates which by default are the spatial
locations of the data (U,V) and the time index (T).

\code{true_cpts}  - The locations of the true changepoints in the dataset.

\code{sp_seg_cng} - The type of spatial change between changepoints (i.e. for each segment)

\code{t_seg_cng}  - The type of temporal change between changepoints (i.e. for each segment)
}
\description{
This functions generates a simulated dataset based on the specified scenario. Users are able to
generate simulated datasets for various changes in the spatial, temporal or spatio-temporal structure
of the dataset. See the GAM-PELT paper for more details.
}
\examples{

#Generate a simulation dataset.
#Set the seed for reproducibility
#100 timesteps with true cpt at 50.
#50 locations.
#Scenario 4b has change is both the spatial and temporal components of the time series.
ds_sim <- gen_SimData('4b',1234,n_ts=100,n_sites=50,true_cpts=c(50))
}
\author{
Michael Hollaway

Rebecca Killick
}
