% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gampelt.R
\name{ncpts}
\alias{ncpts}
\alias{ncpts-methods}
\alias{ncpts,GAM_PELT-method}
\alias{ncpts,GAM_PELT,ncpts-methods}
\title{Generic Function - ncpts}
\usage{
ncpts(object)

\S4method{ncpts}{GAM_PELT}(object)
}
\arguments{
\item{object}{Object of class \code{GAM_PELT}}
}
\value{
The total number of cpts identified.
}
\description{
Generic Function
}
\details{
Generic Function
}
\examples{
#Generate a new object of class GAM_PELT
x <- methods::new('GAM_PELT')
#Retrieve the ncpts slot from x.
ncpts(x)

}
\author{
Michael Hollaway

Rebecca Killick
}
