% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gampelt.R
\name{param.est}
\alias{param.est}
\alias{param.est-methods}
\alias{param.est,GAM_PELT-method}
\alias{param.est,GAM_PELT,param.est-methods}
\title{Generic Function - param.est}
\usage{
param.est(object)

\S4method{param.est}{GAM_PELT}(object)
}
\arguments{
\item{object}{Object of class \code{GAM_PELT}}
}
\value{
A list where each element is a vector of GAM model parameter estimates.
}
\description{
Generic Function
}
\details{
Generic Function
}
\examples{
#Generate a new object of class GAM_PELT
x <- methods::new('GAM_PELT')
#Retrieve the param.est slot from x.
param.est(x)

}
\author{
Michael Hollaway

Rebecca Killick
}
