% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelt.R
\name{PELT}
\alias{PELT}
\title{PELT (Pruned Exact Linear Time)}
\usage{
PELT(df_in, n, minseglen, GAM_call, pen.value, verbose)
}
\arguments{
\item{df_in}{A data frame containing the data within which you want to find a changepoint. Currently assumes all variables to fit the formula set in GAM_call are present.}

\item{n}{The number of time points in the dataset}

\item{minseglen}{Positive Integer giving the minimum segment length (number of observations between changes).}

\item{GAM_call}{The call to provide to the \code{gam} function from \code{mgcv}}

\item{pen.value}{Penalty value, this should be set elsewhere as a numerical value}

\item{verbose}{Logical, print verbose progress output if TRUE}
}
\value{
A list is returned with elements:

\code{lastchangecpts} - Vector of length n containing the last changepoint prior to each timepoint

\code{checklist} - The PELT checklist

\code{cpt_locs} - Ordered list of the optimal number of changepoints

\code{lastchangelike} - Vector of length n containing the likelihood of the optimal segmentation up to each timepoint
}
\description{
Implements the PELT method for identifying changepoints in a given set of summary statistics for a specified cost function and penalty.
This function is called by \code{GAM.PELT}.  This is not intended for use by regular users of the package.  It is exported for developers to call directly for speed increases or to fit alternative cost functions.

WARNING: No checks on arguments are performed!
}
\references{
PELT Algorithm: Killick R, Fearnhead P, Eckley IA (2012) Optimal detection of changepoints with a linear computational cost, \emph{JASA} \bold{107(500)}, 1590--1598
}
\author{
Michael Hollaway

Rebecca Killick
}
