% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gampelt.R
\name{pen.value}
\alias{pen.value}
\alias{pen.value-methods}
\alias{pen.value,GAM_PELT-method}
\alias{pen.value,GAM_PELT,pen.value-methods}
\title{Generic Function - pen.value}
\usage{
pen.value(object)

\S4method{pen.value}{GAM_PELT}(object)
}
\arguments{
\item{object}{Object of class \code{GAM_PELT}}
}
\value{
The value of the penalty used in the analysis.
}
\description{
Generic Function
}
\details{
Generic Function
}
\examples{
#Generate a new object of class GAM_PELT
x <- methods::new('GAM_PELT')
#Retrieve the pen.value slot from x.
pen.value(x)

}
\author{
Michael Hollaway

Rebecca Killick
}
