% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty_decision.R
\name{penalty_decision}
\alias{penalty_decision}
\title{Penalty Decision Function}
\usage{
penalty_decision(penalty, n, diffparam)
}
\arguments{
\item{penalty}{Choice of "BIC/SIC", "AIC" or "MBIC". The predefined penalties listed do count the changepoint as a parameter.}

\item{n}{The length of the original data}

\item{diffparam}{The difference in the number of parameters (degrees of freedom) when a change is added. Changepoint is automatically added so should not be included in this number.}
}
\value{
The numeric value of the penalty
}
\description{
Evaluates the arguments to return a numeric value for the penalty. This function is called by \code{GAM.PELT}.
}
\details{
This function takes the text string input and converts it to a numerical value for the specific length of data specified by n.
}
\author{
Michael Hollaway

Rebecca Killick
}
